Allow Histogram to have a singular bin.
Currently, Histogram assumes that its underflowBin !== overflowBin, and that
neither of them appears in centralBins.
centralBins could be empty, so Histogram can currently have no fewer than 2
bins.

In order to convert TBMv1 Values to Histograms, we don't want to block on
assigning correct bins to all TBMv1 metrics, and we don't want to assume default
bins which can be misleading, so Histogram must be able to contain only a
single bin.

We'd called this "binless mode", but it isn't really a separate mode, nor is it
bin-less. These will be "single-bin Histograms".

This CL implements support for singular bins by removing underflowBin,
overflowBin, and centralBins. There is only allBins.

HistogramBinBoundaries.build_ takes over the responsibility of ensuring that
there are underflow and overflow bins when the nominal range doesn't cover
-Number.MAX_VALUE and Number.MAX_VALUE. The binRanges for the underflow and
overflow bins work just like the other binRanges; they aren't special.

HistogramBinBoundaries.SINGULAR is a static HBB instance that produces a single
binRange. The chart-json converter can pass this HBB to Histogram().

In a followup CL, for single-bin histograms, histogram-span will display a
box-and-whisker plot instead of a NameBarChart with a single useless bar.

BUG=catapult:#2956

Review-Url: https://codereview.chromium.org/2452613003